avoid TB if PR is against the upstream repo
Before this commit:

if a PR is a branch within upstream fork,
then the branch already has the upstream url as remote origin,
and `remote?` returns true. Therefore we don't create an upstream
ref.

After this commit: For now, punt and always add the upstream url.
Git handles this fairly gracefully if the url is already there.