ci: make it easier to run build/test scripts locally
First, move commands out of circle config into standalone scripts.

Second, rename script/* as ci/*.
This means we have to remove the "friction" gem and unit test
since it hard-codes the script dir as "script".

It bothers me to put stuff in "script/*" since that interferes
with tab-completion (cf the "script" command).